[TMP] website_product_filters: This css and less not fine here.
This css refer to a custom set of css necessary just when wbesite_less
is present and refer to place holders into the specific theme where this
module will be used into production, then either move the css to the
correct place (if necessary) and delete the commented code.

The TMP tag is because it can not be necesaary if yes, please delete the
files less and css also, if not uncomment and refer to a correct asset
place this is installing an incorrect set of colors (not the default
one).